Change type of `gap_per_desk` from `u16` to `f32`

+++ ixwindow/wms/i3/src/config.rs
@@ -14,7 +14,7 @@ pub struct Config {
     pub config_dir: String,
     pub cache_dir: String,
     pub color: String,
-    pub gap_per_desk: u16,
+    pub gap_per_desk: f32,
 }
 
 impl Config {
